[DIA] New dot-product based fragmentation matching
The tool MS-DIAL (Tsugawa et. al., 2015) describes a good algorithm
for matching fragmentation spectra. It combines a weighted
dot-product, reverse weighted dot-product and the ratio of actual
to expected number of matches to give a robust score. It manages to
give enough siginificance to both the fragmentation profile's trend
and absolute number of matches.