[Fix] Consistent isotope distribution for integrated groups #1225
Peak-groups that have already been integrated and added to a table
are meant to be frozen in terms of their contents and properties.
However, isotope widget always pulls isotopes at current global
settings, which the user could have changed after doing detection
or manual integration. Selected a group then, would show a
different set of isotopes being populated in isotope widget than
the ones shown in peak table (under the parent peak-group). This
has been fixed and the isotope widget will now show the same set
of isotopologues that are actually present in a parent group.