[Fix] Terminate crashserver on normal app exit #1295
The crashserver daemon has been hooked to only respond to crash
dump requests until now. In case there is no crash event and
El-MAVEN exits normally, the daemon continues to listen in a limbo
state. The Breakpad library, however, does provide for listening to
client exit events along with crash events. A handler for this has
now been installed to take care of the same.