[UI] Add a settings log for individual peak-groups
Since each peak-group will now have a `MavenParameters` object of
its own, there is a need to inspect at what settings a group was
integrated. The new "Settings Log" widget serves this purpose. It
can be accessed from the context menu of single peak-groups or the
new icon for the same, in peak tables.