[Ftr] Edit additional settings in peak-editor
A desirable feature has been to be able to "link" isotopic peak
ranges. This has been available in the isotope dialog and takes
effect while creation of isotopes. But isotopes RT ranges (just
like for other peak-groups) can be edited and again there was a
need to link elution range. This control will now be available in
the peak-editor while editing groups that belong to an isotopologue
family. Furthermore, if the group being edited is the parent
isotope and it was orignally created as "linked" to its other
isotopic forms, then the checkbox will be enabled by default.

Similarly, different peak-groups may require different baseline
estimation parameters. The controls for the same have been added
to the peak-editor.