[Ref] Use reference counting for shared peak-groups
Peak-groups that originate from a peak-table, EIC or from any of
the other widgets in the main-window are usually created and passed
around as raw pointers. The reason behind this is that we want to
maintain uniqueness and allow for identification of potentially
large structures. But too often these groups would get invalidated
or garbage collected while another widget or maybe another thread
was still using them - which would naturally lead to crashes.

To deal with this issue, managed (shared) pointers to peak-groups
will now be used for storing these objects (in containers as well
as state), and wherever possible they will be passed as such. If a
method that takes a peak-group reference as an argument, it should
be passed a raw pointer only if it immediately returns or if it
makes a copy of the peak-group object for its own further use.