[Fix] Incorrect baseline for RT constrained slices
Baselines estimated while pulling EICs for peaks that are expected
to be in a certain RT range were calculated over only that range.
This will not give the same baseline as when it would have been
calculated over the entire chromatogram for that m/z, since both
the baseline estimation algorithms have a somewhat "global" nature
and their results depends on the content of the entire signal. To
make this consistent compared to non-RT restricted peak search,
all EICs will now be extracted over the full RT range for each
sample and then reduced to the desired range after its baseline is
computed.