[Ftr] Detection of independent isotopologues and adducts
El-MAVEN (and MAVEN) has always relied on the presence on parent
(unlabelled) form of a metabolite before attempting to find any
labelled isotopologues. Similar requirements were imposed on adduct
detection as well (and rightly so).

But it is sometimes desirable, to be able to search for labelled
forms of a metabolite, where the parent may not exist at all (for
e.g., injected standards). A similar case can be made for adducts
as well. This required a considerable amount of change to the
detection logic as well as some data-restructuring.

Note: as of now, the IsotopeDetection and AdductDetection module
have not been removed from the codebase, for reference purposes.
But, they will be removed once the new solutions are proven to work
well.