[Ftr] Option to mandate presence of parent ion
Knowing which parent ions of metabolites (out of the entire list
of possible metabolites) are already present in any of the samples,
drastically reduces the search space and improves performance of
detection against a database significantly. Therefore, apart from
being able to search for any arbitrary parent/non-parent ion, a
new option (checkbox) has been added to both isotope and adduct
detection settings, which allow the user to mandate the presence of
a metabolite's normal form before attempting to search for their
labelled isotopologues (or even non-parent adducts).