[Ftr] Allow drag-n-drop among compound's children
Sometimes, automated detection is not able to pick up certain
isotopologues that should have been added to a particular parent
group. In these cases, the user has to manually integrate these
groups and then merge them into the peak-table of interest where it
may still not be attributed as the child of some parent peak-group.
To help with this scenario, a drag-and-drop system has been added
that allows dragging child groups to parent groups that represent
the same compound. This change will be reflected internally as well
as in any table exports.