[Opt] Prevent unnecesary (and expensive) calls
Peak-grouping had reportedly slowed down compared to earlier
(v0.10.0 and before) releases. The primary reason for this was a
large map being constructed and copied for each `MavenParameters`
object which is itself created for each peak-group. As a result
`EIC::groupPeaks` method was often producing the heaviest stack
trace in a typical call to `PeakDetector::processSlices` method.
This copy operation was removed since the map only has a practical
use-case for the global parameters object.

Additionally, a call to `PeakGroup::getFragmentationEvents` would
search for MS2 scans even when there are no MS2 scans in a given
sample for each peak. A check has been placed to prevent a possibly
expensive search for scans.