[Fix] Incorrect m/z calculation for EI type
The calculation of m/z differs only for EI ionization type. This
was not taken into account when calculating the m/z value from
parent adduct ions (which should still obey the EI/non-EI rules).

Additinally, the logic for setting ionization type in the
`MavenParameters` class was constantly assigning ESI even when
unintended.