[Fix] RT deviation abused when checking overlap #1351
When the user opts to "Consider overlap" in peak grouping, the
condition should make the grouping criteria more strict. However,
the logic for group scoring was written such that peaks were
allowed to be grouped if either RT deviation was not too bad or
group overlap was not too bad (which made it more lenient). This
has been fixed.