[Fix] Compound mz value when saved project is loaded
When a project is saved no charge mz is saved. While loading the
same project, mz is calculated according to charge. This leads to
making of a new compound with charged mz and hence we get suffix
in compounds name.
This has been corrected by setting mz with no charge state.